requirements.txt

azure-core==1.35.1
azure-data-tables==12.7.0
azure-functions==1.23.0
azure-identity==1.25.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.1
idna==3.10
isodate==0.7.2
MarkupSafe==3.0.2
msal==1.33.0
msal-extensions==1.3.1
multidict==6.6.4
propcache==0.3.2
pycparser==2.23
PyJWT==2.10.1
requests==2.32.5
six==1.17.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.20.1
pytest 
pytest-mock

